Split up scraper into separate files, to improve maintainability
